Create show-page-name
Bookmark any page in your browser, then add the bookmark to your bookmarks toolbar. Change the URL of your bookmark to the code above. Next time you are in the Adobe Reports & Analytics pages report just click the bookmark to display the default URL associated to each page name in the report.